Call yyparse() + add comments to json_parser.{l,y}
The parser now uses yyparse(). Note that it's very easy to trigger an
error which I specifically note in the comments. I've made it very clear
that this is incomplete, it's easy to trigger an error etc. This is not
even a first draft; it's just a starting point so I can do more as I
learn more about the tools.