Minor updates to JSON parser
Enable debug output in JSON parser. In json_parser.y:

	#define YYDEBUG 1
	int yydebug = 1;

Changed the printing of (in json_parser.l) whitespace, strings and
numbers. Surround strings with 's so that it's more distinguishable as
to what the string is (which includes the "s). For numbers I removed the
text "(as string)" and surrounded the text with 's as well. For
whitespace I surrounded them by 's as well (was "s).

Run make json_parser.