JSON parser now reads mkiocccentry json files OK
The parser now reads the .info.json and .author.json files generated by
mkiocccentry without any errors. This does NOT mean that the parser does
anything! All it means is that the parser can read the files to EOF
without encountering any errors.

The parser grammar is incomplete still and the parser still has no
actions.

The lexer has had some actions updates. First it now uses strdup() on
strings. It doesn't check for errors and the parser does not free()
anything yet (this is noted in comments). Second is whitespace is no
longer returned to the parser. This is because it causes a lot of
problems with either unexpected whitespace token (if it's not in the
respective rules) or expected whitespace (if it is in the rule and there
are no other options); this allows for the parser to read all of the
files now and whitespace is still allowed. Third is I merged JSON_TRUE
and JSON_FALSE into one token JSON_BOOLEAN (JSON_BOOL was already taken
in json.h and they might end up being merged in the end); the action
uses string_to_bool in util.c. Fourth and last is I renamed JSON_EQUALS
to JSON_COLON (the logic behind the previous name is "test_mode":true is
saying that test_mode is equal to true but it might be confusing to some
so I changed it). This change also fixed another problem in that before
I had referenced JSON_COLON in json_parser.y but in json_parser.l it was
JSON_EQUALS which caused an error.

The parser has had some changes as well. First remove references to
JSON_WHITESPACE. Second the JSON_TRUE and JSON_FALSE have been merged to
reflect the lexer as just JSON_BOOLEAN. Third is yyerror() now shows how
many errors had been encountered. Fourth is I enabled debug information
(error messages are also more detailed).

Makefile updates: split json_parser rule into two. To generate the
source files run make parser; to compile the parser run make json_parser
which depends on parser. The parser rule does not depend on any files
right now as I want to be able to always generate new files even if I
haven't made any changes.

Remove json_parser.yy.c as this was generated by flex but I changed it
so flex writes to json_parser.c (which was accidentally added yesterday)
instead. I might change it so that bison writes to this (so bison
creates json_parser.{h,c}) but this will require some other minor
changes as well and can come at a later date.  All of the filenames are
subject to change anyway.