json_parser reads files and strings like jparse
This commit is to set up what I expect to be the next commit namely that
I want json_parser to actually be jparse because that is the goal of
jparse in the first place: to parse strings and files as json. This
commit does the following.

json_parser now includes jparse code but instead of just writing the
file or string to stdout it uses yyparse() on it. It does not create a
tree yet and it probably won't for a while but this way once that's done
all I'll have to do is add the code that verifies the json is correct.

The grammar is still incomplete and in the parser there are no actions
to the rules yet. I'm aware of (due to incomplete grammar) some things
that pass as okay that are not actually okay (in addition to ones I was
already aware of). There might be other things that have to change as
well but this is a good starting point.

A note about the yyparse(). For files it does not read it in as a
string first. Perhaps it should but for now it does not. For strings it
creates a buffer and if != NULL it runs yyparse(); for files if it can
be opened it uses yyparse() on the file (after using yyrestart()).

Makefile updated as json_parser now uses sanity.c which means it has to
link in sanity.o, utf8_posix_map.o and location.o.