json_parse -> jparse
In order for this to work every time I modify jparse.y I have to run
make parser. This is an unfortunate problem caused by the fact that
yacc (I use bison but at the command line running `make' after updating
jparse.y it wants to use yacc for some reason) wants to write to
jparse.c so it does:

	mv -f y.tab.c jparse.c

...which overwrites jparse.c which is from flex. The solution is to make
flex write to a different file and let bison write to jparse.c (which is
probably ideal anyway) but this will require a bit more thought and time
which I do not have right now. It seems there might be a built-in rule
in make that is causing this but I've not investigated this yet.

I would make the jparse rule depend on parser but this would not work
for systems without flex and bison (or possibly different versions).
This too is a work in progress! What this means is I just have to always
run make parser but this should not be a problem because I always run
make before commit and if I get any compilation errors I do not commit.