Disable built-in make rules for .y/.l files
This means that jparse.c won't be overwritten when modifying jparse.y
and running make without running `make parser'`. I still have to run
`make parser' every time I update the files but to help remind myself I
also have jparse recompiled every time these files are changed (they
don't need them to compile but they depend on them so that the tool is
recompiled if they change and since these files and the source/header
files are in the repo this should be fine). Eventually the reliance on
these tools should no longer be or will be limited and this won't be
much of a problem but this way everyone can compile the code as long as
I regenerate the file after I update the lexer/parser files.