Minor update (comments included) in jparse lexer
Remove '.' match in grammar as this is not needed and it only showed
more unnecessary output and did not even return any type to the parser.

Update comment about the parser not being written to be say not finished
as it's not finished but the parser is in a sense written. This is
unimportant but I try keeping comments up to date.