lexer grammar fixes and updates
Separate floating point and integers. This involves a new type in the
union and changing the regexps for the number types. There's a new
comment about the numbers and how to address the type given to simplify
the rules in the parser but this will be worked out later.

Fix the regexp for the floating point/exponent. At least some of it's
been fixed; there might be more cases I have missed but looking at
syntax highlighting in vim I have caught several invalid number strings.
Another minor adjustment there was from a misreading when adding
(somewhat hastily and possibly I was extra tired without realising it)
the original grammar. Specifically remove '""?' from the front. I'm not
sure if there was any harm in having it but it seems unnecessary anyway.

Add an int token_type in the parser for (for now at least) the
punctuation like '{', '}', '[', ']', ':', ',' etc. The assignment occurs
in the lexer actions and currently is not used. It might be useful to
change this to be a type of variable in the union (perhaps with changing
how it's assigned) perhaps in a struct but this will be figured out on a
later date.

Add commented out option in jparse.l that might be of use later but for
now is a problem ('%option nodefault').