parse_json_file(): check for embedded NUL bytes
This was not originally done due to a problem after reading in the file
but without giving it any more thought. Using rewind() on yyin solves
the problem.

Sequenced exit codes. This of course does not work on the *.l or *.y
files but since the code is copied verbatim to the C files this should
be okay. Eventually though the embedded NUL byte error will be a JSON
specific error using jerr() so this won't be an issue anyway.