Fix sequenced exit codes
There actually is an issue with the seqcexit tool not acting on
lexer/parser files (if it does do that I don't know - I've not actually
looked) as it leaves the *.l/*.y files different than the *.c files.
I've manually updated the jparse.l file to match the jparse.c file
(after regenerating the jparse.c).