Convert two errors in jparse.l to warnings
The errors in question are when the file couldn't be read in (i.e.
read_all() returns NULL) and if there are embedded NUL byte(s) in the
file.

I also increment num_errors in in these cases as well as some other
warnings that did not have this done already. This allows for the parser
to continue to the next file or string to actually make testing easier.

Sequenced exit codes (which did not end up needing to modify any other
files but it was done anyway).