Minor bug fixes in json scanner
When parse_json_file() cannot read in the file or if it's not a C string
we use the new function clearerr_or_fclose() which depending on the file
it calls clearerr() or fclose().

Updated Makefile to use seqcexit on the flex and bison files and
sequenced exit codes. I'm aware of an issue though: although it appears
that the same exit code stays in jparse.l as jparse.c there happens to
be a call to err() in mkiocccentry.c with the same error code.