Remove toolkit version, change txzchk -f to -T
As discussed with Landon at
https://github.com/ioccc-src/mkiocccentry/issues/132#issuecomment-1079967380
and
https://github.com/ioccc-src/mkiocccentry/issues/132#issuecomment-1079978436
I have removed the -T option from all the tools. This includes usage
messages, from all the getopt() calls and the man pages.

Because of this I've changed back the txzchk option -f to -T which is
slightly more clear as to what it's purpose is - the file is a text file
rather than just any file (as -f might suggest).