Remove IOCCC_TOOLKIT_RELEASE from Makefile
I thought of this but by the time it came to committing it had slipped
my mind again. Since the toolkit version has been removed from the tools
it no longer needs to be referenced in the Makefile either since it's
not needed in limit_ioccc.sh. It wouldn't have hurt anything but it
needn't be there either.