JSON reserved warn/err codes now 0 - 199
The reason 200 codes are reserved is because it's more than enough ever
and we don't want to have to ever change the codes after the parser and
checkers are complete as this would cause problems for the tools as well
as the test suites. Previously the range was 0 - 99 and although this is
also probably more than enough we want to be sure that there is never a
problem and we cannot imagine how 200 codes will ever not be a large
enough range but if the use of reserved codes change this might not be
the case for just 100 (unlikely though that is).