Rename json warn/err code ignore related names
Where a name was in the form of '.*_code_set' prepend 'json_' to 'code'.
The below I have given specific notes on but this list does not include
all the name changes.

* add_ignore_code() -> ignore_json_code()

The reason I removed the add_ is because the name 'ignore_json_code()'
flows more naturally when reading it.

* is_code_ignored() -> is_json_code_ignored()

Perhaps this should also be renamed in a similar way to
ignore_json_code() only 'json_code_ignored()' e.g.:

    if (json_code_ignored()) { /* ... */ }

but for now this is not the case.

I believe that includes all the name changes that are worth noting but
if I've missed one or more I'm happy to explain it later.

There were also some typos: alloc_ was prepended to the table in some
comments but the table does not have alloc_ as part of it;
alloc_json_code_ignore_set() does (was alloc_code_ignore_set()) but the
table does not.