Move some structs about in json.h
Moved the following structs closer to the top of the file:

    struct integer

    struct floating

    struct string

This is because it's quite possible that the structs will be used in a
new struct/union for the parser so this prevents the need of having to
forward declare them. It might be that this struct/union (yet to be
created) is put in jparse.h instead but this is not yet decided.

As well I've added some notes about the older json_field and json_value
structs as well as the JSON_ type constants (e.g. JSON_STRING, JSON_NUM
etc.).