Minor clean up parser/lexer + make seqcexit
The functions clearerr_or_fclose() and print_newline() have been moved
to util.c so that the definitions are not needed in the lexer.

The print_newline() function now takes a bool output_newline. I'm aware
of another file that prints a newline based on a bool but that file does
not use this routine. Whether util.c is the best file I don't know and
I'm more than fine with it elsewhere.

Sequenced exit codes.