Set yyin = NULL after fclose()
This is done in parse_json_file(). Although yyin is only referenced in
this function if this changes and another function references it it
would be an invalid pointer. The parse_json_file() sets yyin to stdin or
a file associated with the path depending on the path ("-" == stdin) but
this should have been done in the first place to be safer.