Added BISON_VERSION and FLEX_VERSION
Added BISON_VERSION and FLEX_VERSION to version.h, and thus they
will also appear in version.sh.

Added minimum bison version needed to build the JSON parser.
The bison version is based on this command:

	bison -V 2>&1 | head -n 1 | awk '{print $NF}'

Currently BISON_VERSION is: 3.8.2

Added minimum flex version needed to build the JSON parser.
The flex version is based on this command:

	flex -V 2>&1 | head -n 1 | awk '{print $NF}'

Currently FLEX_VERSION is: 2.6.4

This change is in support of the future bisonflexok.sh command.