Change verge to ignore beyond whitespace in version string
After trimming any leading whitespace, the next whitespace
and beyond is trimmed.  This change will allow verge to operate
on this repo's style version strings such as:

    ./verge '0.4 2022-03-31' '0.1 2022-04-02'

While verge is designed to be used by bisonflexok.sh,
this change will allow verge to determine if any repo
version string of the form "a.b YYYY-MM-DD" to be treated
as if it were just "a.b", should that be needed.