Ignore JSON code set updates
Added JSON_CODE_INVALID which is defined as:

    #define JSON_CODE_INVALID (JSON_CODE_RESERVED_MIN - 1)

which is used in place of -1 when initialising the code set. Currently
this is -1 because JSON_CODE_RESERVED_MIN is 0 but if for some strange
reason that value changes this will make it so the invalid value is
correct (the reserved min - 1). Also this removes the magic number -1
which by itself is a worthy change.

IGNORE_JSON_CODE_CHUNK -> JSON_CODE_IGNORE_CHUNK. The rationale here is
that all the other json code constants are prefixed with JSON_CODE_ so
that it's easier to find. MIN_JSON_CODE reads more naturally than
JSON_CODE_MIN but this makes the name scheme consistent: all constants
start with JSON_CODE so that there can be no doubt that it's specific to
JSON codes.