Fix issue where make seqcexit updated parser/lexer
If one were to do:

    make seqcexit
    make seqcexit

They would see something like:

    $ make seqcexit
    seqcexit -c -- jparse.l jparse.y
    seqcexit -- utf8_posix_map.c jparse.c jparse.tab.c dbg.c util.c mkiocccentry.c iocccsize.c fnamchk.c txzchk.c jauthchk.c jinfochk.c json.c jstrencode.c jstrdecode.c rule_count.c location.c sanity.c utf8_test.c jint.c jint.test.c jfloat.c jfloat.test.c verge.c

    $ make seqcexit
    rm -f jparse.tab.h jparse.tab.c
    /opt/local/bin/bison -d jparse.y
    rm -f jparse.c
    /opt/local/bin/flex -o jparse.c jparse.l
    seqcexit -c -- jparse.l jparse.y
    seqcexit -- utf8_posix_map.c jparse.c jparse.tab.c dbg.c util.c mkiocccentry.c iocccsize.c fnamchk.c txzchk.c jauthchk.c jinfochk.c json.c jstrencode.c jstrdecode.c rule_count.c location.c sanity.c utf8_test.c jint.c jint.test.c jfloat.c jfloat.test.c verge.c

What happened? There are two parts to this one of which I can explain
certainly and the other I am pretty sure to be the case but I'm not
absolutely certain. First of all the seqcexit tool updates the files
including the lexer and parser source files.

Now because the make seqcexit rule depended on the source files
including jparse.c _which itself depends on the lexer and parser files_
(which have just been updated) the files had to be regenerated. I believe
this might be because of commit 54722313 and in particular:

        Removed the built-in rules for both ".c <== .y" and ".c <== .l".

but I am not sure of this because the reason I disabled those did not
actually occur after the second run of make seqcexit. But the fact that
make seqcexit always runs seqcexit on the files even if the files have
not changed means this fix should be fine.

The fix is removing the dependency of

    ${ALL_CSRC} ${FLEXFILES} ${BISONFILES}

from make seqcexit (Makefile is still a dependency though).

As for the problem with the running make seqcexit: Most people probably
wouldn't and I'm not even sure how I discovered it but because it would
regenerate the files not everything was updated as make parser use_ref
was not done.

Interestingly running:

    make seqcexit seqcexit

did not cause the problem: it would run seqcexit once and then it would
say seqcexit is up to date.