Fix issue where make seqcexit updated parser/lexer
If one were to do:

    make seqcexit
    make seqcexit

They would see something like:

    $ make seqcexit
    seqcexit -c -- jparse.l jparse.y
    seqcexit -- utf8_posix_map.c jparse.c jparse.tab.c dbg.c util.c mkiocccentry.c iocccsize.c fnamchk.c txzchk.c jauthchk.c jinfochk.c json.c jstrencode.c jstrdecode.c rule_count.c location.c sanity.c utf8_test.c jint.c jint.test.c jfloat.c jfloat.test.c verge.c

    $ make seqcexit
    rm -f jparse.tab.h jparse.tab.c
    /opt/local/bin/bison -d jparse.y
    rm -f jparse.c
    /opt/local/bin/flex -o jparse.c jparse.l
    seqcexit -c -- jparse.l jparse.y
    seqcexit -- utf8_posix_map.c jparse.c jparse.tab.c dbg.c util.c mkiocccentry.c iocccsize.c fnamchk.c txzchk.c jauthchk.c jinfochk.c json.c jstrencode.c jstrdecode.c rule_count.c location.c sanity.c utf8_test.c jint.c jint.test.c jfloat.c jfloat.test.c verge.c

What happened? There are two parts to this one of which I can explain
certainly and the other I believe to be the case but I'm not absolutely
certain. First of all the seqcexit tool updates the files including the
lexer and parser source files. Then (and this is the part I'm not 100
per cent certain of) the fact the lexer and parser source files (*.l and
*.y) have been touched (timestamp) and the fact they are a dependency of
some rules means they needed to be updated. But I am entirely uncertain
as to why it happened without running make jparse. It might be the
change from commit 54722313 and in particular:

        Removed the built-in rules for both ".c <== .y" and ".c <== .l".

but I am not sure of this. I could test adding those back but since
the second invocation of make seqcexit did not cause the problem that
those built-in rules were meant to fix and since make seqcexit runs even
when no files have been modified (despite the fact the rule had the
source files as dependencies) I think this will suffice.

The fix is removing the dependency of

    ${ALL_CSRC} ${FLEXFILES} ${BISONFILES}

from make seqcexit (Makefile is still a dependency though).

As for the problem with the running make seqcexit: Most people probably
wouldn't and I'm not even sure how I discovered it but because it would
regenerate the files not everything was updated as make parser use_ref
was not done.