Added some XXX notes to Makefile wrt sorry.tm.ca.h
The following problems exist (that I'm aware of) which cause the
programmer's apology to be removed from some files. I've added these
comments to the Makefile as it's easier to see where the problems still
exist.

Doing

    make parser

will remove some the programmer's apology from some of the files. But
doing make parser use_ref will solve it (so will make prep).

Also doing

    make seqcexit

will upon recompilation remove the apology from some of the files.

The workaround for now is to make sure you have all the tools installed
and then do make prep or otherwise do what I included in the comments.
These issues will be fixed so that these should not be necessary and at
that point the comments can be removed.