Typo fixes in jparse.y
Removed duplicate words in comments.

Slightly reword why jerr() cannot be used (but a variant could) in
yyerror(). This comment is for later when an error function is actually
used - if one is even used (it might not be necessary at all; this
will be decided later on probably after the parser is finished or when
it has been developed enough where it might be useful).