txzchk: Fix bug introduced in commit 76bf1ab
This commit caused a NULL pointer dereference which was never detected
because the code:

    if (dir_name == NULL || !strlen(dir_name)) {
    -   err(26, __func__, "txzchk: %s: unexpected NULL pointer from %s", txzpath, cmd);
    +   err(26, __func__, "txzchk: unexpected NULL pointer from fnamchk -- %s", txzpath);

was never reached. Specifically in this commit:

    Changed txzchk.c to use shell_cmd() and pipe_open() interfaces

    Modified txzchk.c to use the new cshell_cmd() and pipe_open() interfaces.

The problem was that the char *cmd was no longer used which means that
it remained NULL from the point of initialisation. That's good but as of
that commit it was no longer used so I've removed it entirely.