Add non-fatal error functions werr() and werrp()
The purpose of these functions is to indicate that something is an error
but not make it a fatal error. The logic behind this is it's more clear
that it's supposed to be an error rather than just a warning but
nonetheless the program can continue. This is useful for when a program
(such as jparse) wants to process more than one file or string but be
able to indicate an error rather than a warning (yyerror()
notwithstanding). There are other places where warn() is used instead of
err() exactly because we don't want to abort but we want it to be an
error.

Right now it takes an error_code but I'm not sure if this is desired or
not; if it is perhaps the seqcexit tool should be updated. Otherwise it
doesn't need to be updated and the error_code could be removed (or some
other system could be devised). Given that seqcexit is for exit codes
maybe it shouldn't be there at all but this is probably something to be
discussed. This is one of the reasons the functions are not yet used
(it'll also take some time to discover all the places that this is used
though some immediately come to mind).

The original thought was that this would be for json as there are places
in code where I use warn() simply because I don't want the program to
abort despite the fact it's actually an error. However it occurred to me
that maybe there should be a set of functions of this form for json as
well with not as much information given (as the parser cannot know all
the things that the jwarn()/jerr() functions expect though they could be
modified to not print those out if they're a certain value).

For both sets I'm unsure if the functions should use different sets of
error codes or not and that's another reason they're not used yet -
though the json counterparts of these functions haven't even been
written.

This is what the functions look like:

/*
 * werr - issue an error message as a warning
 *
 * given:
 *	error_code	error code
 *	name		name of function issuing the error
 *	fmt		format of the warning
 *	...		optional format args
 *
 * Example:
 *
 *	werr(1, __func__, "bad foobar: %s", message);
 *
 * NOTE: We warn with extra newlines to help internal fault messages stand out.
 *	 Normally one should NOT include newlines in warn messages.
 */
void
werr(int error_code, char const *name, char const *fmt, ...)

/*
 * werrp - issue an error message with errno information
 *
 * given:
 *	error_code	error code
 *	name		name of function issuing the warning
 *	fmt		format of the warning
 *	...		optional format args
 *
 * Example:
 *
 *	werrp(1, __func__, "bad foobar: %s", message);
 *
 * NOTE: We warn with extra newlines to help internal fault messages stand out.
 *	 Normally one should NOT include newlines in warn messages.
 *
 */
void
werrp(int error_code, char const *name, char const *fmt, ...)