Add extension override option to fnamchk
This option -E ext is strictly for testing and specifically for testing
of txzchk in case I (or we) ever have to add a new tarball listing
format or otherwise for us to test possible (even if improbable)
listings from contestant abuse.

The reason this is useful is because we can type up (or modify a
listing) of tarball listings and we can have 'tarballs' (actually text
files) in the repo (under e.g. ./test_txzchk). This prevents having to
include (possibly destructive) tarballs in the repo and it also allows
for editing the listing: in vim if you try editing e.g. test.txz it'll
show you something like:

    ***error*** (tar#Browse) File not readable</Users/luser/code/ioccc/mkiocccentry/test.txz>

if the file does not exist or:

    " tar.vim version v32
    " Browsing tarfile /Users/cody/code/ioccc/xexyl-mkiocccentry/test.txz
    " Select a file with cursor and press ENTER

    gzip: /Users/luser/code/ioccc/mkiocccentry/test.txz: unexpected end of file

if the file does exist but is not a tarball. Obviously directly
modifying a tarball in vim won't work very well either but opening a
text file will. Thus we could then do e.g.:

    ./fnamchk -E txt test.txt

This of course would fail the first check but the idea is that the .txt
extension would now be accepted.

I've updated the man page for this option as well as making some of it
more clear (most of the changes are unrelated to this new option).
Actually much of the man page was rewritten and/or reformatted.

With this commit a test suite for txzchk can now be devised.