Add -E ext option to txzchk (complete 6ba89af)
Yesterday when I added the -E ext option to fnamchk I forgot to update
txzchk to also have the -E ext option so that a test suite for txzchk
can be devised; without this option txzchk will not use fnamchk -E so it
would always expect a txz extension.

As well I added a missing -- in the call to shell_cmd() in
check_tarball().