Update txzchk.1 for commit f7338647
I did not specify how to write a test suite for txzchk here and although
I did on GitHub I'll duplicate it here with more information.

* How to write a test suite for txzchk

First of all we might have a directory for txzchk like for the JSON with
good and bad subdirectories (i.e. test_JSON) e.g. test_txz or
test_txzchk.

Then a script similar to json-test.sh but maybe txz-test.sh or maybe
txzchk-test.sh that can test the good or bad or both files. Now here's
where the magic happens.

With each file in the subdirectories it can end in .txt: this way the
files can be edited with a text editor and we don't have any destructive
tarballs (and if nothing else no binary files added). Here's an example
text file. The text file
entry.12345678-1234-4321-abcd-1234567890ab-2.1649673512.txt with a tar
output listing looks like:

    drwxr-xr-x  0 501    20          0 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/
    -rw-r--r--  0 501    20       1854 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/Makefile
    -rw-r--r--  0 501    20          4 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/extra2
    -rw-r--r--  0 501    20       2815 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/foo
    -rw-r--r--  0 501    20         61 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/prog.c
    -rw-r--r--  0 501    20       2806 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/.author.json
    -rw-r--r--  0 501    20       2869 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/remarks.md
    -rw-r--r--  0 501    20       5235 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/bar
    -rw-r--r--  0 501    20       1498 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/.info.json
    -rw-r--r--  0 501    20          4 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/extra1

Now that that's in place we can run txzchk with both -T and -E txt so
that txzchk does not open a pipe to tar but rather uses fopen(3) on the
file. Here's how it looks:

    $ ./txzchk -T -E txt entry.12345678-1234-4321-abcd-1234567890ab-2.1649673512.txt
    Welcome to txzchk version: 0.10 2022-03-15

    Performing sanity checks on your environment ...
    ... environment looks OK

    Performing checks on tarball ...
    12345678-1234-4321-abcd-1234567890ab-2
    txzchk: entry.12345678-1234-4321-abcd-1234567890ab-2.1649673512.txt size of 946 bytes OK
    drwxr-xr-x  0 501    20          0 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/
    -rw-r--r--  0 501    20       1854 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/Makefile
    -rw-r--r--  0 501    20          4 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/extra2
    -rw-r--r--  0 501    20       2815 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/foo
    -rw-r--r--  0 501    20         61 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/prog.c
    -rw-r--r--  0 501    20       2806 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/.author.json
    -rw-r--r--  0 501    20       2869 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/remarks.md
    -rw-r--r--  0 501    20       5235 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/bar
    -rw-r--r--  0 501    20       1498 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/.info.json
    -rw-r--r--  0 501    20          4 Apr 11 10:38 12345678-1234-4321-abcd-1234567890ab-2/extra1
    txzchk: entry.12345678-1234-4321-abcd-1234567890ab-2.1649673512.txt total size of files 17146 rounded up to 1024 multiple: 17408 OK
    All checks passed.

If -v > 0 more information is shown depending on the verbosity level.

Note that in the shell_cmd() and pipe_open() functions we always specify
-E with the new txzchk global char const *ext which defaults to "txz":
this way we don't have to worry about checking if the user wanted to
override the extension: it's always specified with either "txz" or the
user supplied extension e.g. "txt" (as demonstrated above).

The script then might run through all the files and if good files fail
or if bad files pass there's a problem (it could also run only specific
tests just like the json-test.sh script does).

This feature allows for testing new formats as well: for example if we
discover a tar listing output that does not match macOS/BSD or linux.
This is why I implemented the -T option in the first place though I had
actually thought of it prior to the discovery that linux has a different
format than macOS for the above mentioned reason: it would be easier to
test different tar outputs including those made by someone attempting to
manipulate tar to create invalid and/or funny tarballs (like the IOCCC
Judges did exactly to get a good set of invalid tarballs to test which
was indeed very helpful).