Spelling change in sorry.tm.ca.h
Specifically honor -> honour. There are two reasons but one was only
argued to plead my case of the other (yes this might be a bit
manipulative :( in the mind of some but my reasoning is below and I
think it's more than fair).

The first one is that I'm the one working on the parser and since I use
British English and it also goes against my honour to spell honour honor
:) and in fact the removal of 'u' in American English goes against my
very nature so I think it's fair to keep my honour by spelling it
honour not honor.

The second one is that honor is a US spelling so I argued that to truly
honour Canadians one should actually use their spelling! :)

Of course as a logophile and amateur philologist I will point out that
yes I do know that American English actually keeps some of the original
English more than British English (and maybe Canadian English as well:
not sure) but I'm not going to start a logomachy over a simple spelling
especially when I get my way. :)

Thank you Landon for not objecting to this change - I appreciate it a
lot!