Minor fixes in txzchk
In function check_txz_file() the 'char *p' was unused other than to
check it was != NULL.

In function alloc_txz_file() I renamed the 'p' parameter to 'path' to be
more clear what it was for. The reason it was named 'p' is because
that's what was used for strtok_r() and so I kept it the same name.