Get rid of the #define JSON_ data types
Instead we use the enum element_type with the prefix JTYPE_.

Additionally the JTYPE_ enum values are powers of two. This might not be
strictly necessary: the reason I used it is because currently the
checkers check for json numbers and not specific types so this way I
could do JTYPE_INT|JTYPE_FLOAT. Once a proper parser is complete this
might not be necessary at all - not sure yet how the checker functions
will work.

There still remains the issue that the parser has an enum with the
prefix JSON_ but I cannot control that and it's not yet worked out how
to integrate the two but using that enum is not yet worked out either as
nothing in the parser is worked out yet.

Running make prep passes all tests. Whether this change breaks some of
the working tests in test_JSON or not I do not know: I'm not too worried
about that because the parser isn't written yet and so what matters now
is that make test passes and it does: any files that are not made by
mkiocccentry can be tested later.

As for renaming the JTYPE_ to JSON_ this might be possible but it might
require changing the names of the symbols in jparse.y; I will worry
about that later - or not - but either way it's low priority right now.