Minor fixes to commit 132ed2
In particular I had done a search and replace via sed of 'JSON_NUM' to
be 'JTYPE_INT|JTYPE_FLOAT' but the fields in the tables are all integers
so I've updated the table to just use JTYPE_INT.

As well I've done the same for the case JTYPE_INT|JTYPE_FLOAT in
jinfochk.c, jauthchk.c and json.c and I've added a comment about the
is_number() function and how it does not handle floating points. In this
comment (in all three files) I've noted how one could check for either
type but noted as well that this won't be done in the end because the
parser explicitly extracts the right types based on regular expressions.